pyproject.toml
[project]
name = "raoc-workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "chardet>=7.2.0",
    "keyring>=25.7.0",
    "pdf2docx>=0.5.12",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "psutil>=7.2.2",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-telegram-bot>=22.7",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["raoc*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "reportlab>=4.4.10",
]
